Fix problems caused by list_to_atom.
List to atom could crash due to too long list - system_limit (255 chars).
Also fixes risk of bringing the system down due to too many atoms. Keep
MethodName (and struct member name) as a list if not possible to convert to
an existing atom in order to keep error reporting in estore_server for not
handled methods. (Currently works for us as all atoms will be loaded while
starting the xmlrpc listener.)